Evil-BW16-Web/requirements.txt
flask==2.3.3
flask-socketio==5.3.6
pyserial==3.5
python-engineio==4.7.1
python-socketio==5.9.0
eventlet==0.33.3
